done with the apperance, need to connect to database
